Search Engine with Tools and Agents/requirements.txt
langchain
langchain-core    
ipykernel
python-dotenv
streamlit
beautifulsoup4
langchain_community
faiss-cpu
fastapi
uvicorn
langserve
sse_starlette
arxiv
wikipedia